Create N-Methyldiethanolamine Market.md
The global N-Methyldiethanolamine Market was valued at USD 754.34 million in 2024. It is expected to grow at a Compound Annual Growth Rate (CAGR) of 4.80%, reaching an estimated market value of USD 999.39 million by 2030. The primary factors driving this growth include the increasing demand for MDEA in the oil & gas sector, where it plays a critical role in gas treatment and purification