Create Powder Magnetic Cores Market.md
The global Powder Magnetic Cores Market was valued at US$ 708.85 million in 2023 and is anticipated to reach US$ 2,048.76 million by 2030, witnessing a CAGR of 16.82% during the forecast period 2024-2030.

Powder Magnetic Cores are a type of magnetic component that is used in various electronic devices and systems.These cores are made from a mixture of fine iron particles, called "iron powder," that is molded into a specific shape and size. The iron powder is then sintered, which involves heating the particles to a high temperature to bind them together and create a solid material.