Create N-Formyl Peptide Receptor 2 Market .md
The global N-Formyl Peptide Receptor 2 Market size is projected to reach US$ 140.71 million by 2030, from US$ 50.37 million in 2022, at a CAGR of 16.48% during 2024-2030.

North America N-Formyl Peptide Receptor 2 market size is valued at US$ 28.27 million in 2023, while Europe N-Formyl Peptide Receptor 2 is 15.24 million. The proportion of the North America is 56.12% in 2023, while Europe is 27.22%, and it is predicted that Europe proportion will reach 26.50% in 2030, trailing a CAGR of 16.08% through the analysis period. As for the Europe N-Formyl Peptide Receptor 2 landscape, Germany is projected to reach US$ 8.41 million by 2030 trailing a CAGR of 15.84% during 2024-2030.