Create 4,4'-thiobis-(2-tert-butyl-5-methylphenol) Market.md
The global market for 4,4'-Thiobis-(2-tert-butyl-5-methylphenol) is projected to experience steady growth. The market size in terms of revenue is expected to increase from $6.98 million in 2024 to $10.01 million by 2031, reflecting a CAGR of 5.09% during the forecast period. In terms of volume, the market is projected to grow from 894.0 tons in 2024 to 1,230.0 tons by 2031, with a CAGR of 5.00% between 2025 and 2031.