Create Toy Collectibles Market.md
The Toy Collectibles Market on a Global scale experienced substantial growth, attaining a valuation of US$ 19.92 billion in 2023. Projections indicate a robust trajectory, with an anticipated value of US$ 45.16 billion by 2030. This growth is underpinned by a noteworthy Compound Annual Growth Rate (CAGR) of 10.82% during the forecast period from 2024 to 2030.