Create  luxury sunglasses market.md
The global luxury sunglasses market is seeing steady growth. In 2023, the market was estimated at around USD 11.68 billion and is expected to grow to USD 21.29 billion by 2032, with a CAGR of 6.90% during the forecast period. Increasing high-net-worth individuals, rising awareness related to one's health, alongside the demand for fashion accessories that display an individual's sense of style and social standing are some of the factors driving growth.