Create Mobile Robot Lithium Battery market.md
The global Mobile Robot Lithium Battery market was valued at approximately US$ 366.95 million in 2023. Projections indicate that the market will grow at a CAGR of 7.58% from 2024 to 2030, reaching an estimated value of US$ 598.75 million by 2030. This growth can be attributed to the increasing adoption of automation in industrial and commercial sectors, as well as advancements in battery technology that enhance energy efficiency and performance.